So it look like gnome team decided to remove StageInputMode in commit 8d1ed9f88ac1574588b566b1e483f14ea455d169 for god only knows what reason as the documentation do not mention it.
In order to help his fellow dev, the great gnome team decided to put "messageTray: Port to GrabHelper" as commit message. With such clarity, it is obvious to everybody what the migration path should be and documenting any of this is a comprehensive manner would be a complete waste of time.

After digging for a while, I have to declare myself defeated by the gnome team and simply skip the check altogether, hopping for it to not have dramatic effect.

Once again, gnome didn't failled its reputation by providing a great user interface and backing it up by the most dubious engineering and not documenting any of it.