Remove support for version previous 3.14
We are accumulating cruft, and most of the code is unsteable by now. Gnome team keep thing randomly, like changing org.gnome.shell.overrides button-layout into org.gnome.desktop.wm.preferences button-layout and the most graceful transition path is obviously to crash and restart the whole UI. This change being necessary, having to way to know when the change took place (one must note that the definition or not of the key is not useful, as it won't give you which one is used) I think the only reasonable path forward it simply to reboot this extention and drop support for old versions.