Fix crash on wayland with gnome-shell 3.36.6
For some reasons `Gtk.Settings.get_default()` has null value when we
start new session in gnome-shell 3.36 on wayland. This leads to the
error shown in log as:

TypeError: Gtk.Settings.get_default(...) is null

After this error the addon can not be loaded and unavailable.

This commit adds appropriate checks for Gtk.Settings.get_default()
values.